var/www/intelligence/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
jinja2==3.1.2
email-validator==2.1.0
httpx==0.25.2
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
